% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{lisa}
\alias{lisa}
\title{Compute Moran's Local Index of Spatial Association (LISA) matrix}
\usage{
lisa(
  y,
  x,
  listw,
  comp = NULL,
  zero.policy = FALSE,
  NAOK = FALSE,
  demean = TRUE
)
}
\arguments{
\item{y}{A vector to compute the index relative to.}

\item{x}{A matrix of variables to compute the lisa with y.}

\item{listw}{A listw object specifying the weighting matrix.}

\item{comp}{A vector of integers, strings, or factors indicating broad tissue compartments in which to compute Moran's I.}

\item{zero.policy}{How to deal with zones without neighbors. If TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA}

\item{NAOK}{If TRUE, NAs are removed in computations.}

\item{demean}{If TRUE demean columns of x. Might not want to if x is dummy variables.}
}
\value{
Returns a list of matrices, where each matrix is the multivariate Moran's I for each tissue compartment.
}
\description{
This code was modified from the moran function above. LISA is directional and is the association centered around marker y.
The column means of LISA are equal to Moran's I.
}
