% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{moran}
\alias{moran}
\title{Compute Moran's I matrix}
\usage{
moran(x, listw, comp = NULL, zero.policy = FALSE, NAOK = FALSE)
}
\arguments{
\item{x}{A matrix of variables to compute the bivariate Moran's I}

\item{listw}{A listw object specifying the weighting matrix.}

\item{comp}{A vector of integers, strings, or factors indicating broad tissue compartments in which to compute Moran's I.}

\item{zero.policy}{How to deal with zones without neighbors. If TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA}

\item{NAOK}{If TRUE, NAs are removed in computations.}
}
\value{
Returns a list of matrices, where each matrix is the multivariate Moran's I for each tissue compartment.
}
\description{
This code was modified from the original in the spdep package.
}
